Fix for Trac #6797 - OSX Build Errors w/ SDL and FCollada
Summary:
A user reports two build issues on OS X. An outdate SDL version and missing reference to math library in
FCollada.

Test Plan:
The reporter confirmed that guidance in the ticket addressed his issue with the build. I
indepently reproduced the error and verified that guidence in the ticket addresses the issue.

Reviewers: Itms!

Subscribers: O5 Build system and libraries

Trac Tickets: #6797

Differential Revision: https://code.wildfiregames.com/D5092